Fixed stack overflow in Day 3 part 2
The filter condition was not written correctly. The comparison needs to
be done against the bit. However, the calculation for the answer is
currently wrong. If the vector generation terminates prematurely,
e.g. because the remaining bytes are identical, then it returns the last
bit where they are identical, rather than the remaining bits.